+++ swing/common-model/src/test/java/is/codion/swing/common/model/component/combobox/FilteredComboBoxModelTest.java
@@ -132,7 +132,7 @@ public class FilteredComboBoxModelTest {
 	void testSelection() {
 		AtomicInteger selectionChangedCounter = new AtomicInteger();
 		Consumer<String> selectionListener = selectedItem -> selectionChangedCounter.incrementAndGet();
-		testModel.addSelectionListener(selectionListener);
+		testModel.selectionObserver().addDataListener(selectionListener);
 		testModel.setSelectedItem(BJORN);
 		assertEquals(1, selectionChangedCounter.get());
 		testModel.setSelectedItem(null);
@@ -155,7 +155,7 @@ public class FilteredComboBoxModelTest {
 		testModel.setSelectedItem(SIGGI);
 		testModel.clear();
 		assertEquals(6, selectionChangedCounter.get());
-		testModel.removeSelectionListener(selectionListener);
+		testModel.selectionObserver().removeDataListener(selectionListener);
 	}
 
 	@Test